% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcut.R
\name{bcut}
\alias{bcut}
\title{Cut a vector into intervals with sensible and flexible labels}
\usage{
bcut(x, min = 0, stepsize = 10000, labels = "upper")
}
\arguments{
\item{x}{A numeric vector to be sliced into intervals}

\item{min}{The minimum value to include in the intervals (defaults to zero)}

\item{stepsize}{The size of each interval (defaults to 1000)}

\item{labels}{A string that defines the type of labels will be returned
(defaults to `upper`)}
}
\description{
Cut a vector into intervals with sensible and flexible labels
}
\details{
This is a "better cut" (bcut) wrapper for the base R cut() function,
  which returns factors that describe the intervals that the values of the
  input vector `x` fall into. They're ugly although notationally accurate,
  but we often want to replace them with better, and usually shorter, labels.
  The format of the returned labels are specified in the `labels` parameter.
  If `labels=midranges` the labels will be the midrange value for the
  intervals that `x` is sliced into. If `labels=upper` only the upper end of
  the range for each interval will be returned. If `labels=ranges` the label
  will include both the lower and upper end of the interval, separated by a 
  hyphen. The function intentionally does not cast the returned values as 
  factors, leaving that to the user if desired.
}
\examples{
results$interval <- bcut(results$count, labels = "midpoints")
}
