% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aggregate_error.R
\name{calc_aggregate_error}
\alias{calc_aggregate_error}
\title{Calculate aggregate comparison between two vectors}
\usage{
calc_aggregate_error(df, observed, predicted, places = 1)
}
\arguments{
\item{df}{Name of the tibble (data frame) object containing vectors (columns)
of observed and predicted values}

\item{observed}{A string identifying the column name of the observed values}

\item{predicted}{A string identifying the column name of the predicted values}

\item{places}{An integer denoting the number of places past the decimal that
.  floating point statistics will be reported with. The default value is 1.}
}
\description{
Calculate aggregate comparison between two vectors
}
\details{
This function returns a table containing comparison of two summed
  vectors of data. The totals, the difference, and percent difference
  between them are reported instead of comparisons between individual
  observations within each vector. If the data are grouped these statistics
  will be compiled as expected with a row with totals added to the bottom.
}
\examples{
calc_aggregate_error(assignment_results, "count", "assigned")
# You can group the data when passing it to this function:
calc_aggregate_error(group_by(assignment_results, functional_type),
  "count", "assigned")
}
