% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_NCHRP_statistics.R
\name{calc_NCHRP_statistics}
\alias{calc_NCHRP_statistics}
\title{Calculate the percent of links within MDE and AEC limits}
\usage{
calc_NCHRP_statistics(df, observed, predicted)
}
\arguments{
\item{df}{Name of the tibble (data frame) object containing the data to be 
plotted}

\item{observed}{The variable in `df` containing the counted or observed
flows to compare to}

\item{predicted}{The variable in `df` containing the simulated or 
predicted flows}
}
\description{
Calculate the percent of links within MDE and AEC limits
}
\details{
This function calculates the number of links that fall within 
  the NCHRP limits for maximum desirable error (MDE) and approximate
  error in count (AEC) as defined in NCHRP Report 750.
}
\examples{
eh <- plot_NCHRP_comparison(df, "count", "assigned", "road_type")
print(eh)
pdf(filename, height = 8, width = 11); print(eh); dev.off()
}
