% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_frequency_labels.R
\name{create_frequency_labels}
\alias{create_frequency_labels}
\title{Create frequency labels for a categorical variable}
\usage{
create_frequency_labels(df, varname, suffix = "_lab")
}
\arguments{
\item{df}{Name of the tibble (data frame) object containing the field to be 
be summarised}

\item{varname}{The field containing a string variable for which the
frequencies of its occurrence will be tabulated}

\item{suffix}{The suffix attached to `varname` with the frequency labels
appended to it, defaults to "_lab"}
}
\description{
Create frequency labels for a categorical variable
}
\details{
Categorical variables are often used for colouring or symbols in
  ggplot2 graphics. This function clones a categorical variable and adds
  the frequency in parentheses (n=x) after each value of `varname`. The
  resulting tibble retains `varname`, one record for each value, and a new
  variable that appends the user-supplied (or defaulted to "_lab") suffix
  to `varname`. This tibble can then be joined to the original `df` using 
  `varname` as the by variable. Then use the new variable as the grouping
  variable in ggplot (e.g., colour = varname_lab).
}
\examples{
eh <- create_frequency_labels(df, "Type")
new_df <- left_join(df, eh, by = "Type")
ggplot(new_df, aes(colour = Type_lab, ...)
}
