% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_matrix.R
\name{freq_matrix}
\alias{freq_matrix}
\title{Create two-dimensional matrix of grouped frequencies or weighted sums}
\usage{
freq_matrix(df, group_vars, descending = TRUE, weight_var = NULL)
}
\arguments{
\item{df}{Name of the tibble (data frame) object containing the data to be 
summarised}

\item{group_vars}{A list of two or more variables that will be used to 
group the data by}

\item{descending}{A boolean variable to indicate whether the resulting
tibble will be sorted in descending order by row sum (defaults to TRUE)}

\item{weight_var}{An optional variable number to weight the observations
by when constructing the frequency matrix (defaults to NULL for no 
weighting}
}
\description{
Create two-dimensional matrix of grouped frequencies or weighted sums
}
\details{
This function creates two-dimensional matrix of frequencies or
  weighted frequencies using two or more grouping variables, with the row
  and column totals and percentages appended. It is important to note that
  all of the `group_var`s must be strings to avoid their values being
  unwittingly included in the row and column totals. If using a grouping 
  variable with numeric values cast it to character before passing it to 
  this function.
}
\examples{
eh <- gmatr::freq_matrix(survey_data, c("hhsize", "income"))
}
