% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_matrix.R
\name{gather_matrix}
\alias{gather_matrix}
\title{Convert R matrix to tibble}
\usage{
gather_matrix(matrix, value_name = "value")
}
\arguments{
\item{matrix}{A matrix in R format with row and column names defined}

\item{value_name}{A user-defined column name for the elements of the input
matrix, defaults to `value`}
}
\description{
Convert R matrix to tibble
}
\details{
This is Greg Macfarlane's incredibly fast and efficient code for
  transforming a R matrix to tibble in i, j, value format. You can chain a
  rename() to the return value to rename the row and column names from 
  `origin` and `destination` to whatever names you'd like.
}
\examples{
eh <- gather_matrix(my_matrix, "distance_km")
neh <- gather_matrix(my_matrix) \%>\% rename(key = origin, y = destination)
}
