% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tibble_header.R
\name{get_tibble_header}
\alias{get_tibble_header}
\title{Return a string with the dimensons of a tibble}
\usage{
get_tibble_header(target)
}
\arguments{
\item{df}{Name of the tibble (data frame) object}
}
\description{
Return a string with the dimensons of a tibble
}
\details{
This function simply returns a string identifying the input tibble
  and its dimensions.
}
\examples{
print(get_tibble_header(df))
}
