% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NCHRP_comparison.R
\name{plot_NCHRP_comparison}
\alias{plot_NCHRP_comparison}
\title{Create NCHRP highway validation plot}
\usage{
plot_NCHRP_comparison(
  df,
  observed,
  predicted,
  groupvar,
  plot_title = NULL,
  max_pct_error = 300,
  show_AEC = TRUE,
  show_exclusions = TRUE
)
}
\arguments{
\item{df}{Name of the tibble (data frame) object containing the data to be 
plotted}

\item{observed}{The variable in `df` containing the counted or observed
flows to compare to}

\item{predicted}{The variable in `df` containing the simulated or 
predicted flows}

\item{groupvar}{The variable in `df` to colour the observations by
(often link type or functional classification)}

\item{plot_title}{A character string that corresponds to ggplot's title
label, displayed left-justified at the top of the plot}

\item{max_pct_error}{Integer defining the maximum percent error that will
be shown on the y-axis (see notes below about negative values are
handled), default is 300 percent}

\item{show_AEC}{Boolean variable indicating whether the approximate error
in count (AEC) is shown, defaults to TRUE}

\item{show_exclusions}{Boolean variable indicating whether the number of
observations in `df` that are outside of the `max_pct_error` (and thus,
not shown on the plot) are reported, defaults to TRUE}
}
\description{
Create NCHRP highway validation plot
}
\details{
This function creates a plot of percent error in traffic assignment
  by comparing observed and predicted values as shown in NCHRP Report 750.
  The percent error is shown on the `y` axis and the observed value on the
  `x` axis. The maximum desirable error (MDE) and approximate error in 
  count (AEC) lines are shown, as well as a report of how many observations
  fall within each set of lines. This plot differs from the NCHRP approach
  in that both positive and negative percent errors are shown, rather than
  just absolute values. The observations (points) can be coloured if the 
  user specifies a `groupvar`. If so, the number of observations are 
  reported in the legend beside each level of `groupvar`. The function
  returns a ggplot object that contains the plot with annotations.

  The function normally plots observations with percent error ranging from 
  -150 to 300 percent. The positive value is set by the `max_pct_error`
  parameter. The plot is not symmetrical, but by definition no observation
  can have a negative percent error less than -100 percent. Unless the user
  disables it by setting `show_exclusions` to FALSE an annotation will be 
  added at the top of the plot showing the number of observations not shown
  (i.e., outside of the range specified by `max_pct_error`. If you want to
  show all observations regardless of how compressed the space between the
  MDE lines is simply set `max_pct_error` to the largest percent error in
  your data. (Note that you cannot set `max_pct_error` to a huge number, as
  ggplot will use that as the maximum value on the y-axis.

  The maximum desirable error (MDE) lines (positive and negative) are
  always shown, as that is a key part of the NCHRP display. The approximate
  error in count (AEC) is less often used. It is turned on by default but
  used a dotted line to distinguish it from MDE (solid line). The AEC line
  can be turned off by setting the `show_AEC` parameter to FALSE.
}
\examples{
eh <- plot_NCHRP_comparison(df, "count", "assigned", "road_type")
print(eh)
pdf(filename, height = 8, width = 11); print(eh); dev.off()
}
