% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NCHRP_screenlines.R
\name{plot_NCHRP_screenlines}
\alias{plot_NCHRP_screenlines}
\title{Create NCHRP screenline comparison plot}
\usage{
plot_NCHRP_screenlines(
  df,
  observed,
  predicted,
  screenline,
  plot_title = NULL,
  max_pct_error = 300
)
}
\arguments{
\item{df}{Name of the tibble (data frame) object containing the data to be 
plotted}

\item{observed}{The variable in `df` containing the counted or observed
flows to compare to}

\item{predicted}{The variable in `df` containing the simulated or 
predicted flows}

\item{screenline}{An integer identifier for the screenline that the 
observations are associated with}

\item{plot_title}{A character string that corresponds to ggplot's title
label, displayed left-justified at the top of the plot}

\item{max_pct_error}{Integer defining the maximum percent error that will
be shown on the y-axis (see notes below about negative values are
handled), default is 300 percent}
}
\description{
Create NCHRP screenline comparison plot
}
\details{
This function creates a plot of percent error on screenlines by 
  comparing observed and predicted values as shown in Figure A-9 in NCHRP
  Report 255. The percent error is shown on the `y` axis and the observed 
  values on the `x` axis. The maximum desirable error (MDE) bars are shown,
  as well as a report of how many observations fall between the bounds.
  This plot differs from the NCHRP approach in that both positive and
  negative percent errors are shown, rather than just absolute values. The
  function returns a ggplot object that contains the plot with annotations.

  The function normally plots observations with percent error ranging from 
  -150 to 300 percent. The positive value is set by the `max_pct_error`
  parameter. The plot is not symmetrical, but by definition no observation
  can have a negative percent error less than -100 percent. If you want to
  show all observations regardless of how compressed the space between the
  MDE lines is simply set `max_pct_error` to the largest percent error in
  your data. (Note that you cannot set `max_pct_error` to a huge number, as
  ggplot will use that as the maximum value on the y-axis.
}
\examples{
eh <- plot_NCHRP_screenlines(df, "count", "assigned", "screen_id")
print(eh)
pdf(filename, height = 8, width = 11); print(eh); dev.off()
}
