% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_levels.R
\name{sum_levels}
\alias{sum_levels}
\title{Count the number of levels of a categorical variable}
\usage{
sum_levels(dt, varname, descending = TRUE)
}
\arguments{
\item{dt}{Name of the tibble (data frame) object containing the variable}

\item{varname}{The categorical variable to be summarised}

\item{descending}{A boolean value denoting whether the results should be
sorted in descending order by levels (default is TRUE)}
}
\description{
Count the number of levels of a categorical variable
}
\details{
This function returns the number of observations in each level of a
  categorical variable. The levels are listed in descending frequency by 
  default.
}
\examples{
sum_levels(this_tibble, "this_variable")
}
