% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_numeric.R
\name{sum_numeric}
\alias{sum_numeric}
\title{Tabulate typical summary statistics for continuous numeric variables}
\usage{
sum_numeric(df, var, places = 4, na_rm = TRUE, max_print = 99)
}
\arguments{
\item{df}{Name of the tibble (data frame) object containing the variable}

\item{var}{The numeric variable to be summarised}

\item{places}{An integer specifying the number of places past the decimals to
report the summary statistics (defaults to 4)}

\item{na_rm}{A boolean variable specifying whether missing values will be
ignored in the calculation of the summaries (defaults to TRUE)}

\item{max_print}{The maximum number of rows in the output (defaults to 99)}
}
\description{
Tabulate typical summary statistics for continuous numeric variables
}
\details{
This function returns a tibble with the number of observations, 
  missing values, mean, median, min and max values, and the standard
  deviation for a numeric variable. To summarise the variable by grouping
  variables simply pass the exogenously `group_by` tibble to the `df`
  argument. A row is added to the summary for all levels of the grouping
  variable(s) combined in such cases.
}
\examples{
sum_levels(this_tibble, "this_variable")
sum_levels(group_by(this_tibble, group_var), "this_variable")
}
