% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verified_full_join.R
\name{verified_full_join}
\alias{verified_full_join}
\title{Verify that full_join has observations from both input tibbles}
\usage{
verified_full_join(left_df, right_df, by_var)
}
\arguments{
\item{left_df}{Tibble with one or more columns in common with `right_df`}

\item{right_df}{Tibble with one of more columns in common with `left_db`}

\item{by_var}{A string containing the name(s) of column(s) to join the two
two tibbles by}
}
\description{
Verify that full_join has observations from both input tibbles
}
\details{
This function does a full merge on two tibbles, ensuring that all
  records from both tibbles are included. It then ensures that all records
  are present in both datasets, in which case the joined tibble is returned
  to the calling program. However, if one or more records are present in
  one but not the other the mismatches are displayed and the function
  returns an empty tibble.
}
\examples{
verify_full_join(assigned, counts, by = "link_ID")
}
